fix: fixing files references after renaming
